Script to clear caches and restart database server
This one is rather important for in between benchmark runs.

To minimise the chance that the results are skewed by in-memory caches
from a previous run, this script helps to clear the OS page cache, and
flush any Cassandra memtables to disk, and restarts the database service
itself to clear any other in-memory caches.